HELLCAT STEALER/Settings/requirements.txt
pillow
customtkinter
requests
pyautogui
uuid
pycryptodomex
datetime
colorama
platformdirs
psutil
plyer
pyinstaller